% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ComputeEngine}
\alias{ComputeEngine}
\title{ComputeEngine Object}
\usage{
ComputeEngine(instanceName = NULL, zone = NULL, machineType = NULL,
  diskNames = NULL)
}
\arguments{
\item{instanceName}{The instance on which the operation is running}

\item{zone}{The availability zone in which the instance resides}

\item{machineType}{The machine type of the instance}

\item{diskNames}{The names of the disks that were created for this pipeline}
}
\value{
ComputeEngine object
}
\description{
ComputeEngine Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes a Compute Engine resource that is being managed by a running pipeline.
}

