% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ControllerConfig.gcsSinks}
\alias{ControllerConfig.gcsSinks}
\title{ControllerConfig.gcsSinks Object}
\usage{
ControllerConfig.gcsSinks()
}
\value{
ControllerConfig.gcsSinks object
}
\description{
ControllerConfig.gcsSinks Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ControllerConfig functions: \code{\link{ControllerConfig.disks}},
  \code{\link{ControllerConfig.gcsSources}},
  \code{\link{ControllerConfig.vars}},
  \code{\link{ControllerConfig}}
}

