% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ControllerConfig.gcsSources}
\alias{ControllerConfig.gcsSources}
\title{ControllerConfig.gcsSources Object}
\usage{
ControllerConfig.gcsSources()
}
\value{
ControllerConfig.gcsSources object
}
\description{
ControllerConfig.gcsSources Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ControllerConfig functions: \code{\link{ControllerConfig.disks}},
  \code{\link{ControllerConfig.gcsSinks}},
  \code{\link{ControllerConfig.vars}},
  \code{\link{ControllerConfig}}
}

