% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ControllerConfig}
\alias{ControllerConfig}
\title{ControllerConfig Object}
\usage{
ControllerConfig(ControllerConfig.vars = NULL,
  ControllerConfig.disks = NULL, ControllerConfig.gcsSources = NULL,
  ControllerConfig.gcsSinks = NULL, image = NULL, cmd = NULL,
  gcsLogPath = NULL, machineType = NULL, vars = NULL, disks = NULL,
  gcsSources = NULL, gcsSinks = NULL)
}
\arguments{
\item{ControllerConfig.vars}{The \link{ControllerConfig.vars} object or list of objects}

\item{ControllerConfig.disks}{The \link{ControllerConfig.disks} object or list of objects}

\item{ControllerConfig.gcsSources}{The \link{ControllerConfig.gcsSources} object or list of objects}

\item{ControllerConfig.gcsSinks}{The \link{ControllerConfig.gcsSinks} object or list of objects}

\item{image}{No description}

\item{cmd}{No description}

\item{gcsLogPath}{No description}

\item{machineType}{No description}

\item{vars}{No description}

\item{disks}{No description}

\item{gcsSources}{No description}

\item{gcsSinks}{No description}
}
\value{
ControllerConfig object
}
\description{
ControllerConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Stores the information that the controller will fetch from the server in order to run. Should only be used by VMs created by the Pipelines Service and not by end users.
}
\seealso{
Other ControllerConfig functions: \code{\link{ControllerConfig.disks}},
  \code{\link{ControllerConfig.gcsSinks}},
  \code{\link{ControllerConfig.gcsSources}},
  \code{\link{ControllerConfig.vars}}
}

