% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ControllerConfig.vars}
\alias{ControllerConfig.vars}
\title{ControllerConfig.vars Object}
\usage{
ControllerConfig.vars()
}
\value{
ControllerConfig.vars object
}
\description{
ControllerConfig.vars Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ControllerConfig functions: \code{\link{ControllerConfig.disks}},
  \code{\link{ControllerConfig.gcsSinks}},
  \code{\link{ControllerConfig.gcsSources}},
  \code{\link{ControllerConfig}}
}

