% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Disk}
\alias{Disk}
\title{Disk Object}
\usage{
Disk(name = NULL, type = NULL, sizeGb = NULL, source = NULL,
  autoDelete = NULL, readOnly = NULL, mountPoint = NULL)
}
\arguments{
\item{name}{Required}

\item{type}{Required}

\item{sizeGb}{The size of the disk}

\item{source}{The full or partial URL of the persistent disk to attach}

\item{autoDelete}{Specifies whether or not to delete the disk when the pipeline completes}

\item{readOnly}{Specifies how a sourced-base persistent disk will be mounted}

\item{mountPoint}{Required at create time and cannot be overridden at run time}
}
\value{
Disk object
}
\description{
Disk Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Google Compute Engine disk resource specification.
}

