% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\docType{package}
\name{genomics_googleAuthR}
\alias{genomics_googleAuthR}
\alias{genomics_googleAuthR-package}
\title{Genomics API
Stores, processes, explores and shares genomic data.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:19:23
filename: /Users/mark/dev/R/autoGoogleAPI/googlegenomicsv1alpha2.auto/R/genomics_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/genomics
}
}

