% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ListPipelinesResponse}
\alias{ListPipelinesResponse}
\title{ListPipelinesResponse Object}
\usage{
ListPipelinesResponse(nextPageToken = NULL, pipelines = NULL)
}
\arguments{
\item{nextPageToken}{The token to use to get the next page of results}

\item{pipelines}{The matched pipelines}
}
\value{
ListPipelinesResponse object
}
\description{
ListPipelinesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response of ListPipelines. Contains at most `pageSize`pipelines. If it contains `pageSize` pipelines, and more pipelinesexist, then `nextPageToken` will be populated and should beused as the `pageToken` argument to a subsequent ListPipelinesrequest.
}
