% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{LocalCopy}
\alias{LocalCopy}
\title{LocalCopy Object}
\usage{
LocalCopy(path = NULL, disk = NULL)
}
\arguments{
\item{path}{Required}

\item{disk}{Required}
}
\value{
LocalCopy object
}
\description{
LocalCopy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
LocalCopy defines how a remote file should be copied to and from the VM.
}

