% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{LoggingOptions}
\alias{LoggingOptions}
\title{LoggingOptions Object}
\usage{
LoggingOptions(gcsPath = NULL)
}
\arguments{
\item{gcsPath}{The location in Google Cloud Storage to which the pipeline logs}
}
\value{
LoggingOptions object
}
\description{
LoggingOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The logging options for the pipeline run.
}
