% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Operation.response}
\alias{Operation.response}
\title{Operation.response Object}
\usage{
Operation.response()
}
\value{
Operation.response object
}
\description{
Operation.response Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
If importing ReadGroupSets, an ImportReadGroupSetsResponse is returned. If importing Variants, an ImportVariantsResponse is returned. For pipelines and exports, an empty response is returned.
}
\seealso{
Other Operation functions: \code{\link{Operation.metadata}},
  \code{\link{Operation}}
}
