% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{OperationEvent}
\alias{OperationEvent}
\title{OperationEvent Object}
\usage{
OperationEvent(description = NULL, endTime = NULL, startTime = NULL)
}
\arguments{
\item{description}{Required description of event}

\item{endTime}{Optional time of when event finished}

\item{startTime}{Optional time of when event started}
}
\value{
OperationEvent object
}
\description{
OperationEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An event that occurred during an Operation.
}
