% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{OperationMetadata}
\alias{OperationMetadata}
\title{OperationMetadata Object}
\usage{
OperationMetadata(OperationMetadata.request = NULL,
  OperationMetadata.runtimeMetadata = NULL, OperationMetadata.labels = NULL,
  request = NULL, runtimeMetadata = NULL, createTime = NULL,
  labels = NULL, projectId = NULL, clientId = NULL, events = NULL,
  endTime = NULL, startTime = NULL)
}
\arguments{
\item{OperationMetadata.request}{The \link{OperationMetadata.request} object or list of objects}

\item{OperationMetadata.runtimeMetadata}{The \link{OperationMetadata.runtimeMetadata} object or list of objects}

\item{OperationMetadata.labels}{The \link{OperationMetadata.labels} object or list of objects}

\item{request}{The original request that started the operation}

\item{runtimeMetadata}{Runtime metadata on this Operation}

\item{createTime}{The time at which the job was submitted to the Genomics service}

\item{labels}{Optionally provided by the caller when submitting the request that creates}

\item{projectId}{The Google Cloud Project in which the job is scoped}

\item{clientId}{This field is deprecated}

\item{events}{Optional event messages that were generated during the job's execution}

\item{endTime}{The time at which the job stopped running}

\item{startTime}{The time at which the job began to run}
}
\value{
OperationMetadata object
}
\description{
OperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata describing an Operation.
}
\seealso{
Other OperationMetadata functions: \code{\link{OperationMetadata.labels}},
  \code{\link{OperationMetadata.request}},
  \code{\link{OperationMetadata.runtimeMetadata}}
}
