% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{OperationMetadata.request}
\alias{OperationMetadata.request}
\title{OperationMetadata.request Object}
\usage{
OperationMetadata.request()
}
\value{
OperationMetadata.request object
}
\description{
OperationMetadata.request Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The original request that started the operation. Note that this will be in current version of the API. If the operation was started with v1beta2 API and a GetOperation is performed on v1 API, a v1 request will be returned.
}
\seealso{
Other OperationMetadata functions: \code{\link{OperationMetadata.runtimeMetadata}},
  \code{\link{OperationMetadata}}
}

