% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{operations.cancel}
\alias{operations.cancel}
\title{Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. Clients may use Operations.GetOperation or Operations.ListOperations to check whether the cancellation succeeded or the operation completed despite cancellation.}
\usage{
operations.cancel(CancelOperationRequest, name)
}
\arguments{
\item{CancelOperationRequest}{The \link{CancelOperationRequest} object to pass to this method}

\item{name}{The name of the operation resource to be cancelled}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics/}{Google Documentation}

Other CancelOperationRequest functions: \code{\link{CancelOperationRequest}}
}

