% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{PipelineResources}
\alias{PipelineResources}
\title{PipelineResources Object}
\usage{
PipelineResources(minimumCpuCores = NULL, preemptible = NULL,
  minimumRamGb = NULL, disks = NULL, zones = NULL,
  bootDiskSizeGb = NULL, noAddress = NULL)
}
\arguments{
\item{minimumCpuCores}{The minimum number of cores to use}

\item{preemptible}{Whether to use preemptible VMs}

\item{minimumRamGb}{The minimum amount of RAM to use}

\item{disks}{Disks to attach}

\item{zones}{List of Google Compute Engine availability zones to which resource creation will restricted}

\item{bootDiskSizeGb}{The size of the boot disk}

\item{noAddress}{Whether to assign an external IP to the instance}
}
\value{
PipelineResources object
}
\description{
PipelineResources Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The system resources for the pipeline run.
}

