% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{pipelines.create}
\alias{pipelines.create}
\title{Creates a pipeline that can be run later. Create takes a Pipeline thathas all fields other than `pipelineId` populated, and then returnsthe same pipeline with `pipelineId` populated. This id can be usedto run the pipeline.Caller must have WRITE permission to the project.}
\usage{
pipelines.create(Pipeline)
}
\arguments{
\item{Pipeline}{The \link{Pipeline} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics}{Google Documentation}

Other Pipeline functions: \code{\link{Pipeline}}
}
