% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{pipelines.delete}
\alias{pipelines.delete}
\title{Deletes a pipeline based on ID.Caller must have WRITE permission to the project.}
\usage{
pipelines.delete(pipelineId)
}
\arguments{
\item{pipelineId}{Caller must have WRITE access to the project in which this pipeline}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics}{Google Documentation}
}
