% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{pipelines.get}
\alias{pipelines.get}
\title{Retrieves a pipeline based on ID. Caller must have READ permission to the project.}
\usage{
pipelines.get(pipelineId)
}
\arguments{
\item{pipelineId}{Caller must have READ access to the project in which this pipeline is defined}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics/}{Google Documentation}
}

