% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{pipelines.getControllerConfig}
\alias{pipelines.getControllerConfig}
\title{Gets controller configuration information. Should only be called by VMs created by the Pipelines Service and not by end users.}
\usage{
pipelines.getControllerConfig(operationId = NULL, validationToken = NULL)
}
\arguments{
\item{operationId}{The operation to retrieve controller configuration for}

\item{validationToken}{}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics/}{Google Documentation}
}

