% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{pipelines.run}
\alias{pipelines.run}
\title{Runs a pipeline. If `pipelineId` is specified in the request, thenrun a saved pipeline. If `ephemeralPipeline` is specified, then runthat pipeline once without saving a copy.The caller must have READ permission to the project where the pipelineis stored and WRITE permission to the project where the pipeline will berun, as VMs will be created and storage will be used.If a pipeline operation is still running after 6 days, it will be canceled.}
\usage{
pipelines.run(RunPipelineRequest)
}
\arguments{
\item{RunPipelineRequest}{The \link{RunPipelineRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics}{Google Documentation}

Other RunPipelineRequest functions: \code{\link{RunPipelineRequest}}
}
