% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{pipelines.setOperationStatus}
\alias{pipelines.setOperationStatus}
\title{Sets status of a given operation. Any new timestamps (as determined bydescription) are appended to TimestampEvents. Should only be called by VMscreated by the Pipelines Service and not by end users.}
\usage{
pipelines.setOperationStatus(SetOperationStatusRequest)
}
\arguments{
\item{SetOperationStatusRequest}{The \link{SetOperationStatusRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics}{Google Documentation}

Other SetOperationStatusRequest functions: \code{\link{SetOperationStatusRequest}}
}
