% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{RepeatedString}
\alias{RepeatedString}
\title{RepeatedString Object}
\usage{
RepeatedString(values = NULL)
}
\arguments{
\item{values}{No description}
}
\value{
RepeatedString object
}
\description{
RepeatedString Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
