% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{RunPipelineArgs.inputs}
\alias{RunPipelineArgs.inputs}
\title{RunPipelineArgs.inputs Object}
\usage{
RunPipelineArgs.inputs()
}
\value{
RunPipelineArgs.inputs object
}
\description{
RunPipelineArgs.inputs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Pipeline input arguments; keys are defined in the pipeline documentation. All input parameters that do not have default values must be specified. If parameters with defaults are specified here, the defaults will be overridden.
}
\seealso{
Other RunPipelineArgs functions: \code{\link{RunPipelineArgs.outputs}},
  \code{\link{RunPipelineArgs}}
}

