% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{RunPipelineArgs.labels}
\alias{RunPipelineArgs.labels}
\title{RunPipelineArgs.labels Object}
\usage{
RunPipelineArgs.labels()
}
\value{
RunPipelineArgs.labels object
}
\description{
RunPipelineArgs.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels to apply to this pipeline run. Labels will also be applied tocompute resources (VM, disks) created by this pipeline run. When listingoperations, operations can filtered by labels.Label keys may not be empty; label values may be empty. Non-empty labelsmust be 1-63 characters long, and comply with [RFC1035](https://www.ietf.org/rfc/rfc1035.txt).Specifically, the name must be 1-63 characters long and match the regularexpression `[a-z]([-a-z0-9]*[a-z0-9])?` which means the firstcharacter must be a lowercase letter, and all following characters must bea dash, lowercase letter, or digit, except the last character, which cannotbe a dash.
}
\seealso{
Other RunPipelineArgs functions: \code{\link{RunPipelineArgs.inputs}},
  \code{\link{RunPipelineArgs.outputs}},
  \code{\link{RunPipelineArgs}}
}
