% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{RunPipelineArgs.outputs}
\alias{RunPipelineArgs.outputs}
\title{RunPipelineArgs.outputs Object}
\usage{
RunPipelineArgs.outputs()
}
\value{
RunPipelineArgs.outputs object
}
\description{
RunPipelineArgs.outputs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Pipeline output arguments; keys are defined in the pipelinedocumentation.  All output parameters of without default valuesmust be specified.  If parameters with defaults are specifiedhere, the defaults will be overridden.
}
\seealso{
Other RunPipelineArgs functions: \code{\link{RunPipelineArgs.inputs}},
  \code{\link{RunPipelineArgs.labels}},
  \code{\link{RunPipelineArgs}}
}
