% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{RunPipelineArgs}
\alias{RunPipelineArgs}
\title{RunPipelineArgs Object}
\usage{
RunPipelineArgs(RunPipelineArgs.inputs = NULL,
  RunPipelineArgs.outputs = NULL, projectId = NULL, inputs = NULL,
  outputs = NULL, serviceAccount = NULL, clientId = NULL,
  resources = NULL, logging = NULL, keepVmAliveOnFailureDuration = NULL)
}
\arguments{
\item{RunPipelineArgs.inputs}{The \link{RunPipelineArgs.inputs} object or list of objects}

\item{RunPipelineArgs.outputs}{The \link{RunPipelineArgs.outputs} object or list of objects}

\item{projectId}{Required}

\item{inputs}{Pipeline input arguments; keys are defined in the pipeline documentation}

\item{outputs}{Pipeline output arguments; keys are defined in the pipeline documentation}

\item{serviceAccount}{The Google Cloud Service Account that will be used to access data and services}

\item{clientId}{Client-specified pipeline operation identifier}

\item{resources}{Specifies resource requirements/overrides for the pipeline run}

\item{logging}{Required}

\item{keepVmAliveOnFailureDuration}{How long to keep the VM up after a failure (for example docker command failed, copying input or output files failed, etc)}
}
\value{
RunPipelineArgs object
}
\description{
RunPipelineArgs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The pipeline run arguments.
}
\seealso{
Other RunPipelineArgs functions: \code{\link{RunPipelineArgs.inputs}},
  \code{\link{RunPipelineArgs.outputs}}
}

