% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{RunPipelineRequest}
\alias{RunPipelineRequest}
\title{RunPipelineRequest Object}
\usage{
RunPipelineRequest(pipelineArgs = NULL, pipelineId = NULL,
  ephemeralPipeline = NULL)
}
\arguments{
\item{pipelineArgs}{The arguments to use when running this pipeline}

\item{pipelineId}{The already created pipeline to run}

\item{ephemeralPipeline}{A new pipeline object to run once and then delete}
}
\value{
RunPipelineRequest object
}
\description{
RunPipelineRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request to run a pipeline. If `pipelineId` is specified, itrefers to a saved pipeline created with CreatePipeline and set asthe `pipelineId` of the returned Pipeline object. If`ephemeralPipeline` is specified, that pipeline is run oncewith the given args and not saved. It is an error to specify both`pipelineId` and `ephemeralPipeline`. `pipelineArgs`must be specified.
}
\seealso{
Other RunPipelineRequest functions: \code{\link{pipelines.run}}
}
