% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ServiceAccount}
\alias{ServiceAccount}
\title{ServiceAccount Object}
\usage{
ServiceAccount(email = NULL, scopes = NULL)
}
\arguments{
\item{email}{Email address of the service account}

\item{scopes}{List of scopes to be enabled for this service account on the pipeline virtual machine}
}
\value{
ServiceAccount object
}
\description{
ServiceAccount Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Google Cloud Service Account.
}

