% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SetOperationStatusRequest}
\alias{SetOperationStatusRequest}
\title{SetOperationStatusRequest Object}
\usage{
SetOperationStatusRequest(operationId = NULL, errorMessage = NULL,
  validationToken = NULL, errorCode = NULL, timestampEvents = NULL)
}
\arguments{
\item{operationId}{No description}

\item{errorMessage}{No description}

\item{validationToken}{No description}

\item{errorCode}{No description}

\item{timestampEvents}{No description}
}
\value{
SetOperationStatusRequest object
}
\description{
SetOperationStatusRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to set operation status. Should only be used by VMscreated by the Pipelines Service and not by end users.
}
\seealso{
Other SetOperationStatusRequest functions: \code{\link{pipelines.setOperationStatus}}
}
