% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{TimestampEvent}
\alias{TimestampEvent}
\title{TimestampEvent Object}
\usage{
TimestampEvent(description = NULL, timestamp = NULL)
}
\arguments{
\item{description}{String indicating the type of event}

\item{timestamp}{The time this event occured}
}
\value{
TimestampEvent object
}
\description{
TimestampEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Stores the list of events and times they occured for major events in job execution.
}

