% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Disk}
\alias{Disk}
\title{Disk Object}
\usage{
Disk(autoDelete = NULL, sizeGb = NULL, mountPoint = NULL, source = NULL,
  name = NULL, type = NULL)
}
\arguments{
\item{autoDelete}{Deprecated}

\item{sizeGb}{The size of the disk}

\item{mountPoint}{Required at create time and cannot be overridden at run time}

\item{source}{The full or partial URL of the persistent disk to attach}

\item{name}{Required}

\item{type}{Required}
}
\value{
Disk object
}
\description{
Disk Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Google Compute Engine disk resource specification.
}
