% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{DockerExecutor}
\alias{DockerExecutor}
\title{DockerExecutor Object}
\usage{
DockerExecutor(imageName = NULL, cmd = NULL)
}
\arguments{
\item{imageName}{Required}

\item{cmd}{Required}
}
\value{
DockerExecutor object
}
\description{
DockerExecutor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The Docker execuctor specification.
}

