% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{OperationMetadata}
\alias{OperationMetadata}
\title{OperationMetadata Object}
\usage{
OperationMetadata(OperationMetadata.request = NULL,
  OperationMetadata.runtimeMetadata = NULL, projectId = NULL,
  createTime = NULL, startTime = NULL, endTime = NULL, request = NULL,
  events = NULL, clientId = NULL, runtimeMetadata = NULL)
}
\arguments{
\item{OperationMetadata.request}{The \link{OperationMetadata.request} object or list of objects}

\item{OperationMetadata.runtimeMetadata}{The \link{OperationMetadata.runtimeMetadata} object or list of objects}

\item{projectId}{The Google Cloud Project in which the job is scoped}

\item{createTime}{The time at which the job was submitted to the Genomics service}

\item{startTime}{The time at which the job began to run}

\item{endTime}{The time at which the job stopped running}

\item{request}{The original request that started the operation}

\item{events}{Optional event messages that were generated during the job's execution}

\item{clientId}{Optionally provided by the caller when submitting the request that creates the operation}

\item{runtimeMetadata}{Runtime metadata on this Operation}
}
\value{
OperationMetadata object
}
\description{
OperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata describing an Operation.
}
\seealso{
Other OperationMetadata functions: \code{\link{OperationMetadata.request}},
  \code{\link{OperationMetadata.runtimeMetadata}}
}

