% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{PipelineResources}
\alias{PipelineResources}
\title{PipelineResources Object}
\usage{
PipelineResources(bootDiskSizeGb = NULL, minimumRamGb = NULL,
  preemptible = NULL, zones = NULL, minimumCpuCores = NULL,
  noAddress = NULL, disks = NULL)
}
\arguments{
\item{bootDiskSizeGb}{The size of the boot disk}

\item{minimumRamGb}{The minimum amount of RAM to use}

\item{preemptible}{Whether to use preemptible VMs}

\item{zones}{List of Google Compute Engine availability zones to which resource}

\item{minimumCpuCores}{The minimum number of cores to use}

\item{noAddress}{Whether to assign an external IP to the instance}

\item{disks}{Disks to attach}
}
\value{
PipelineResources object
}
\description{
PipelineResources Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The system resources for the pipeline run.
}
