% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{pipelines.list}
\alias{pipelines.list}
\title{Lists pipelines.Caller must have READ permission to the project.}
\usage{
pipelines.list(namePrefix = NULL, pageToken = NULL, pageSize = NULL,
  projectId = NULL)
}
\arguments{
\item{namePrefix}{Pipelines with names that match this prefix should be}

\item{pageToken}{Token to use to indicate where to start getting results}

\item{pageSize}{Number of pipelines to return at once}

\item{projectId}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/genomics}{Google Documentation}
}
