% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{RunPipelineArgs}
\alias{RunPipelineArgs}
\title{RunPipelineArgs Object}
\usage{
RunPipelineArgs(RunPipelineArgs.labels = NULL,
  RunPipelineArgs.outputs = NULL, RunPipelineArgs.inputs = NULL,
  logging = NULL, labels = NULL, keepVmAliveOnFailureDuration = NULL,
  resources = NULL, outputs = NULL, projectId = NULL, clientId = NULL,
  serviceAccount = NULL, inputs = NULL)
}
\arguments{
\item{RunPipelineArgs.labels}{The \link{RunPipelineArgs.labels} object or list of objects}

\item{RunPipelineArgs.outputs}{The \link{RunPipelineArgs.outputs} object or list of objects}

\item{RunPipelineArgs.inputs}{The \link{RunPipelineArgs.inputs} object or list of objects}

\item{logging}{Required}

\item{labels}{Labels to apply to this pipeline run}

\item{keepVmAliveOnFailureDuration}{How long to keep the VM up after a failure (for example docker command}

\item{resources}{Specifies resource requirements/overrides for the pipeline run}

\item{outputs}{Pipeline output arguments; keys are defined in the pipeline}

\item{projectId}{Required}

\item{clientId}{This field is deprecated}

\item{serviceAccount}{The Google Cloud Service Account that will be used to access data and}

\item{inputs}{Pipeline input arguments; keys are defined in the pipeline documentation}
}
\value{
RunPipelineArgs object
}
\description{
RunPipelineArgs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The pipeline run arguments.
}
\seealso{
Other RunPipelineArgs functions: \code{\link{RunPipelineArgs.inputs}},
  \code{\link{RunPipelineArgs.labels}},
  \code{\link{RunPipelineArgs.outputs}}
}
