% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ServiceAccount}
\alias{ServiceAccount}
\title{Genomics API Objects 
Upload, process, query, and search Genomics data in the cloud.}
\usage{
ServiceAccount(email = NULL, scopes = NULL)
}
\arguments{
\item{email}{Email address of the service account}

\item{scopes}{List of scopes to be enabled for this service account on the VM}
}
\value{
ServiceAccount object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:53:15
filename: /Users/mark/dev/R/autoGoogleAPI/googlegenomicsv1alpha2.auto/R/genomics_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ServiceAccount Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Google Cloud Service Account.
}
